portalicious: handle mutation errors in confirmation dialog
AB#30503

This refactor makes the assumption that confirmation dialogs will always be used with mutations, which I think is a fair assumption.

If we find a scenario where this isn't true, we can refactor / create an abstraction later.

+++ interfaces/Portalicious/src/app/domains/project/project.api.service.ts
@@ -157,7 +157,11 @@ export class ProjectApiService extends DomainApiService {
     );
   }
 
-  removeProjectUser(projectId: Signal<number>, userId: number) {
+  removeProjectUser(projectId: Signal<number>, userId?: number) {
+    if (!userId) {
+      return Promise.reject(new Error('User ID is required'));
+    }
+
     return this.httpWrapperService.perform121ServiceRequest<Project>({
       method: 'DELETE',
       endpoint: `${BASE_ENDPOINT}/${projectId().toString()}/users/${userId.toString()}`,
